+++ spec/examples/usual/dynamic_options/format/date/properties/pattern/example3_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
+RSpec.describe Usual::DynamicOptions::Format::Date::Properties::Pattern::Example3 do
   describe ".call!" do
     subject(:perform) { described_class.call!(**attributes) }
 
@@ -14,7 +14,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[started_on],
+                     internals: %i[],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -23,7 +23,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
           result = perform
 
           expect(result.started_on?).to be(true)
-          expect(result.started_on).to eq(Date.parse(started_on))
+          expect(result.started_on).to eq("2023/04/14")
         end
       end
 
@@ -34,9 +34,9 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Internal,
-                "[Usual::DynamicOptions::Format::Date::Options::Example2] " \
-                "Internal attribute `started_on` does not match `date` format"
+                ApplicationService::Exceptions::Output,
+                "[Usual::DynamicOptions::Format::Date::Properties::Pattern::Example3] " \
+                "Output attribute `started_on` does not match `date` format"
               )
             )
           end
@@ -66,7 +66,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[started_on],
+                     internals: %i[],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -75,7 +75,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
           result = perform
 
           expect(result.started_on?).to be(true)
-          expect(result.started_on).to eq(Date.parse(started_on))
+          expect(result.started_on).to eq("2023/04/14")
         end
       end
 
@@ -86,9 +86,9 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Options::Example2 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Internal,
-                "[Usual::DynamicOptions::Format::Date::Options::Example2] " \
-                "Internal attribute `started_on` does not match `date` format"
+                ApplicationService::Exceptions::Output,
+                "[Usual::DynamicOptions::Format::Date::Properties::Pattern::Example3] " \
+                "Output attribute `started_on` does not match `date` format"
               )
             )
           end
